refactor(game): move archetypes to game dir
Rename `estado.*` to `state.*`.
+ Move archetypes `state.*`, `tree.*`, `obstacle.*` and `player.*` to `ecs/systems/`. + fix imports.
