+++ src/game/obstacle.c
