+++ src/game/state.c
@@ -1,3 +1,3 @@
-#include "estado.h"
+#include "state.h"
 
 Modo modoAtual = MODO_MENU;
