+++ src/game/state.h
@@ -1,8 +1,5 @@
-#ifndef ESTADO_H
-#define ESTADO_H
+#pragma once
 
 typedef enum { MODO_MENU, MODO_JOGO, MODO_GAMEOVER } Modo;
 
 extern Modo modoAtual;
-
-#endif
